backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
web3>=6.0.0
eth-account>=0.11.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
apscheduler>=3.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
